refactor(include)!: put the headers in directory utility within include.
The goal of this change is to formalize this scheme across my libraries.
Previously, I didn't think this was that important. However, thinking
about it, this is actually quite important. It puts the headers of this
library into their own scope so that when you have multiple libaries,
this minimizes the chances of getting name collision between files.

BREAKING_CHANGE headers are now within direcotyr `include/utility/`.
Previously, they were in directory `include/`.