refactor: refactor the general `toString` function
Why?
The original code implemented two cuasi-exact functions, with the same
signature except in the constraint requirments. This created a problem
for documentation, as well as, it wasn't elegant.

How?
Use constexpr if statements for compiler-dependant conditional compilat-
ion. This is a fascinating new feature.
Also, add a new concept to test for more types that support string conv-
ersion, namely, those that do so via a `stringfy` functin. The concept
is named `Stringfyble`.