docs: finish the documentation
Why?

The documentation is absolutely necessary. You know this when you create
a project and then come to use it or modify it a few months later.

How?

Write the documentation using reStructuredText and Sphinx.

Others:

Add a version and a description to the project.

Add a new concept "HasContains".

Refactor two of the `toString` function overloads. This function had
two overloads for different kinds of iterables, based on whether they
had a certain operation or not. Using conditional compilation, the two
overloads have been combined into one.